Added theming for chessboard
