+++ client/src/components/ChessBoardConfigs/ChessBoardConfigs.css
@@ -49,3 +49,283 @@
 .bluePossibleMoveMark {
     background-color: #1565C0;
 }
+
+.magicLightSquare {
+    background-image: url(../../assets/pieces/magicPieces/square-light.png);
+    background-size: cover;
+    opacity: .8;
+    filter: grayscale(100%);
+}
+
+.magicDarkSquare {
+    background-image: url(../../assets/pieces/magicPieces/square-dark.png);
+    background-size: cover;
+    opacity: .8;
+    filter: grayscale(70%);
+}
+
+.hideFigureEffectWithLight {
+  position: relative;
+  overflow: visible;
+  animation: lightningStrike 0.6s ease-out forwards;
+  transform-origin: center center;
+}
+
+.hideFigureEffectWithLight::before {
+  content: '';
+  position: absolute;
+  top: -100%;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 4px;
+  height: 200%;
+  background: linear-gradient(
+    to bottom,
+    transparent 0%,
+    rgba(255, 255, 255, 0.8) 10%,
+    rgba(135, 206, 250, 0.9) 20%,
+    rgba(255, 255, 255, 1) 30%,
+    rgba(135, 206, 250, 0.9) 40%,
+    rgba(255, 255, 255, 0.8) 50%,
+    transparent 100%
+  );
+  box-shadow: 
+    0 0 20px rgba(255, 255, 255, 0.8),
+    0 0 40px rgba(135, 206, 250, 0.6),
+    0 0 60px rgba(255, 255, 255, 0.4),
+    -8px 0 10px rgba(255, 255, 255, 0.3),
+    8px 0 10px rgba(255, 255, 255, 0.3);
+  animation: lightningBolt 0.3s ease-out forwards;
+  z-index: 1000;
+}
+
+.hideFigureEffectWithLight::after {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: radial-gradient(
+    circle at center,
+    rgba(255, 255, 255, 0.9) 0%,
+    rgba(135, 206, 250, 0.6) 30%,
+    transparent 70%
+  );
+  animation: lightningFlash 0.35s ease-out forwards;
+  z-index: 999;
+}
+
+.hideFigureEffectWithBurn {
+  position: relative;
+  overflow: visible;
+  animation: burnEffect 0.8s ease-out forwards;
+  transform-origin: center center;
+}
+
+.hideFigureEffectWithBurn::before {
+  content: '';
+  position: absolute;
+  bottom: -20px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 80px;
+  height: 60px;
+  background: radial-gradient(
+    ellipse at center,
+    rgba(255, 100, 0, 0.9) 0%,
+    rgba(255, 150, 0, 0.7) 20%,
+    rgba(255, 200, 0, 0.5) 40%,
+    rgba(255, 100, 0, 0.3) 60%,
+    transparent 100%
+  );
+  border-radius: 50% 50% 50% 50% / 60% 60% 40% 40%;
+  box-shadow: 
+    0 0 30px rgba(255, 100, 0, 0.8),
+    0 0 60px rgba(255, 150, 0, 0.6),
+    0 0 90px rgba(255, 200, 0, 0.4),
+    inset 0 -10px 20px rgba(255, 50, 0, 0.5);
+  animation: flameRise 0.8s ease-out forwards;
+  z-index: 1000;
+}
+
+.hideFigureEffectWithBurn::after {
+  content: '';
+  position: absolute;
+  bottom: 20px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 40px;
+  height: 100px;
+  background: linear-gradient(
+    to top,
+    rgba(100, 100, 100, 0.8) 0%,
+    rgba(80, 80, 80, 0.6) 30%,
+    rgba(60, 60, 60, 0.4) 60%,
+    rgba(40, 40, 40, 0.2) 80%,
+    transparent 100%
+  );
+  border-radius: 20px;
+  filter: blur(8px);
+  animation: smokeRise 0.8s ease-out forwards;
+  z-index: 999;
+}
+
+@keyframes burnEffect {
+  0% {
+    opacity: 1;
+    transform: scale(1) translateY(0);
+    filter: brightness(1) hue-rotate(0deg);
+  }
+  20% {
+    opacity: 1;
+    transform: scale(1.05) translateY(-2px);
+    filter: brightness(1.3) hue-rotate(10deg);
+  }
+  40% {
+    opacity: 0.95;
+    transform: scale(1.1) translateY(-5px);
+    filter: brightness(1.5) hue-rotate(20deg) saturate(1.5);
+  }
+  60% {
+    opacity: 0.7;
+    transform: scale(1.15) translateY(-10px);
+    filter: brightness(1.8) hue-rotate(30deg) saturate(2);
+  }
+  80% {
+    opacity: 0.4;
+    transform: scale(1.2) translateY(-20px);
+    filter: brightness(2) hue-rotate(40deg) saturate(2.5);
+  }
+  100% {
+    opacity: 0;
+    transform: scale(0.8) translateY(-40px);
+    filter: brightness(1) hue-rotate(0deg) grayscale(0.8);
+  }
+}
+
+@keyframes flameRise {
+  0% {
+    width: 60px;
+    height: 40px;
+    bottom: -10px;
+    opacity: 0.8;
+  }
+  30% {
+    width: 100px;
+    height: 70px;
+    bottom: -30px;
+    opacity: 1;
+  }
+  60% {
+    width: 120px;
+    height: 90px;
+    bottom: -50px;
+    opacity: 0.9;
+  }
+  100% {
+    width: 80px;
+    height: 50px;
+    bottom: -60px;
+    opacity: 0;
+  }
+}
+
+@keyframes smokeRise {
+  0% {
+    opacity: 0;
+    transform: translateX(-50%) translateY(0) scale(0.8);
+  }
+  20% {
+    opacity: 0.6;
+    transform: translateX(-50%) translateY(-20px) scale(1);
+  }
+  50% {
+    opacity: 0.8;
+    transform: translateX(-50%) translateY(-50px) scale(1.2);
+  }
+  80% {
+    opacity: 0.5;
+    transform: translateX(-50%) translateY(-80px) scale(1.5);
+  }
+  100% {
+    opacity: 0;
+    transform: translateX(-50%) translateY(-120px) scale(1.8);
+  }
+}
+
+@keyframes lightningStrike {
+  0% {
+    opacity: 1;
+    transform: translate(0, 0) rotate(0deg) scale(1);
+    filter: brightness(1);
+  }
+  20% {
+    opacity: 1;
+    transform: translate(0, 0) rotate(0deg) scale(1.05);
+    filter: brightness(2.5);
+  }
+  30% {
+    opacity: 1;
+    transform: translate(10px, -5px) rotate(5deg) scale(1.1);
+    filter: brightness(3);
+  }
+  50% {
+    opacity: 0.9;
+    transform: translate(50px, -20px) rotate(25deg) scale(1.15);
+    filter: brightness(2);
+  }
+  70% {
+    opacity: 0.6;
+    transform: translate(120px, -40px) rotate(45deg) scale(1.1);
+    filter: brightness(1.5);
+  }
+  100% {
+    opacity: 0;
+    transform: translate(200px, -80px) rotate(90deg) scale(0.8);
+    filter: brightness(1);
+  }
+}
+
+@keyframes lightningBolt {
+  0% {
+    top: -100%;
+    opacity: 0;
+  }
+  15% {
+    opacity: 1;
+  }
+  30% {
+    top: 0%;
+    opacity: 1;
+  }
+  35% {
+    opacity: 0.9;
+  }
+  50% {
+    top: 50%;
+    opacity: 0.5;
+  }
+  100% {
+    top: 100%;
+    opacity: 0;
+  }
+}
+
+@keyframes lightningFlash {
+  0% {
+    opacity: 0;
+  }
+  15% {
+    opacity: 1;
+  }
+  30% {
+    opacity: 0.8;
+  }
+  50% {
+    opacity: 0.3;
+  }
+  100% {
+    opacity: 0;
+  }
+}
